pyproject.toml
[tool.poetry]
name = "tradingproject"
version = "0.1.0"
description = ""
authors = ["KingBrado <berardo.m.manzi@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
streamlit = "^1.40.2"
plotly = "^5.24.1"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.3.1"
pandas = "^2.2.3"
yfinance = "^0.2.50"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
